Transform options keys to symbols.
Before this, any options with a String key (e.g. "host") instead of a Symbol key (e.g. :host) would get silently ignored.

This is surprising in general but in particular, when using Rails and reading the `database.yml` config, they are all String keys so have no affect unless they are transformed before passing to `Trilogy.new`.

This reduces the surprises and ensures the keys are all transformed to Symbols before accessing them.